+++ lnd/index.js
@@ -3,6 +3,7 @@ const findRecord = require('./find_record');
 const gateway = require('./gateway');
 const getCertValidityDays = require('./get_cert_validity_days');
 const getCredentials = require('./get_credentials');
+const getLnds = require('./get_lnds');
 const lndCredentials = require('./lnd_credentials');
 
 module.exports = {
@@ -11,5 +12,6 @@ module.exports = {
   gateway,
   getCertValidityDays,
   getCredentials,
+  getLnds,
   lndCredentials,
 };